refactor: JdbcTemplate singleton 변경
