+++ nextstep-jdbc/src/main/java/nextstep/jdbc/JdbcTemplate.java
@@ -1,5 +1,6 @@
 package nextstep.jdbc;
 
+import javax.sql.DataSource;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -9,8 +10,22 @@ import java.util.List;
 import java.util.Optional;
 
 public class JdbcTemplate {
+    private static JdbcTemplate jdbcTemplate;
+    private DataSource dataSource;
+
+    private JdbcTemplate(DataSource dataSource) {
+        this.dataSource = dataSource;
+    }
+
+    public static JdbcTemplate getInstance(DataSource dataSource) {
+        if (jdbcTemplate == null) {
+            return new JdbcTemplate(dataSource);
+        }
+        return jdbcTemplate;
+    }
+
     public void executeUpdate(String sql, Object... queryParams) {
-        try (Connection con = ConnectionManager.getConnection();
+        try (Connection con = dataSource.getConnection();
              PreparedStatement pstmt = con.prepareStatement(sql)) {
             setQueryParams(pstmt, queryParams);
             pstmt.executeUpdate();
@@ -20,7 +35,7 @@ public class JdbcTemplate {
     }
 
     public <T> List<T> selectAll(String sql, RowMapper rowMapper, Object... queryParams) {
-        try (Connection con = ConnectionManager.getConnection();
+        try (Connection con = dataSource.getConnection();
              PreparedStatement pstmt = con.prepareStatement(sql)) {
 
             setQueryParams(pstmt, queryParams);
@@ -39,7 +54,7 @@ public class JdbcTemplate {
 
     public <T> T selectObject(String sql, RowMapper rowMapper, Object... queryParams) {
         Optional<T> result = Optional.empty();
-        try (Connection con = ConnectionManager.getConnection();
+        try (Connection con = dataSource.getConnection();
              PreparedStatement pstmt = con.prepareStatement(sql)) {
 
             setQueryParams(pstmt, queryParams);
