+++ slipp/src/main/java/slipp/controller/UpdateUserController.java
@@ -1,6 +1,5 @@
 package slipp.controller;
 
-import nextstep.jdbc.JdbcTemplate;
 import nextstep.mvc.asis.Controller;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -14,7 +13,7 @@ import javax.servlet.http.HttpServletResponse;
 public class UpdateUserController implements Controller {
     private static final Logger log = LoggerFactory.getLogger(UpdateUserController.class);
 
-    private final UserDao userDao = new UserDao(new JdbcTemplate());
+    private final UserDao userDao = new UserDao();
 
     @Override
     public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
